Changes to bidijkstra

+++ OSM_Unity_Client/Assets/Tests/editmode/biDijkstraTest.cs
@@ -14,7 +14,7 @@ public class NewTestScript
         var denmarkGraph = MapController.DeserializeGraph("Assets/Maps/denmark.graph");
         long startNode = 280276691;
         long endNode = 896780523;
-        double expectedDistance = 324741;
+        float expectedDistance = 324741;
 
         BiDijkstra biDijkstra = new BiDijkstra(denmarkGraph);
         var pathResult = biDijkstra.FindShortestPath(startNode, endNode);
